Add combineOpenGroups rx-operator
